+++ packages/plugin-html/src/toolbar/createToolbar.js
@@ -1,9 +1,9 @@
 import createInlineButtons from './inlineButtons';
 import createInsertButtons from './insertButtons';
 
-export default function createToolbar({ helpers, t, isMobile }) {
+export default function createToolbar({ helpers, t, isMobile, settings }) {
   return {
-    InlineButtons: createInlineButtons({ t }),
+    InlineButtons: createInlineButtons({ t, settings }),
     InsertButtons: isMobile ? [] : createInsertButtons({ helpers, t }),
     name: 'html',
   };